+++ lib/routes/japanpost/namespace.ts
@@ -1,6 +1,12 @@
 import type { Namespace } from '@/types';
 
 export const namespace: Namespace = {
-    name: 'Japanpost 日本郵便',
+    name: 'Japanpost',
     url: 'trackings.post.japanpost.jp',
+    zh: {
+        name: '日本邮政',
+    },
+    ja: {
+        name: '日本郵便',
+    },
 };
